Render proficiency signal icon on desktop module cards
Extracts the signal icon logic from ModuleRow into a shared
ProficiencySignal component (used identically by both breakpoints,
differing only in the monochrome color token) and renders it in
ModuleCard's bottom row beside the "Completed" label, black/50 to
match this card's existing tokens. Renders only for completed modules
with a non-null proficiency score.

Refs #333
