+++ app/language-learning/level/[level]/components/ModuleRow.tsx
@@ -4,7 +4,7 @@ import { RoundButton } from 'toto-react';
 import { ModuleProgressEntry } from '@/api/TomeLearningDashboardAPI';
 import { ProgressBar } from '@/app/ui/general/ProgressBar';
 import { MaskedSvgIcon } from '@/app/components/MaskedSvgIcon';
-import { getProficiencyLevel } from '@/utils/proficiencyLevel';
+import { ProficiencySignal } from '@/app/components/ProficiencySignal';
 
 const STEP_NUMBER: Record<string, number> = {
     grammar: 1,
@@ -12,28 +12,6 @@ const STEP_NUMBER: Record<string, number> = {
     test: 3,
     done: 3,
 };
-
-const PROFICIENCY_ICON: Record<1 | 2 | 3 | 4, string> = {
-    1: '/images/signal-weak.svg',
-    2: '/images/signal-fair.svg',
-    3: '/images/signal-good.svg',
-    4: '/images/signal.svg',
-};
-
-function ProficiencySignal({ proficiency }: { proficiency: ModuleProgressEntry['proficiency'] }) {
-    if (proficiency === null) return null;
-
-    const level = getProficiencyLevel(proficiency.score);
-
-    return (
-        <MaskedSvgIcon
-            src={PROFICIENCY_ICON[level]}
-            alt={`Proficiency level ${level} of 4`}
-            color="bg-black/70"
-            backgroundSrc="/images/signal.svg"
-        />
-    );
-}
 
 function ModuleNode({ status, num }: { status: ModuleProgressEntry['status']; num: string }) {
     const isHighlighted = status === 'in_progress' || status === 'completed';
@@ -136,7 +114,7 @@ export function ModuleRow({
                 />
             )}
 
-            {module.status === 'completed' && <ProficiencySignal proficiency={module.proficiency} />}
+            {module.status === 'completed' && <ProficiencySignal proficiency={module.proficiency} color="bg-black/70" />}
         </div>
     );
 }
